Merge remote-tracking branch 'upstream/combined' into combined

+++ src/daysim/sun.h
@@ -12,7 +12,7 @@ double stadj( int jd);                 /* solar time adjustment from Julian date
 double sdec( int jd);                  /* solar declination angle from Julian date */
 float  f_altitude(float lat,float sd,float st);
 float  f_azimuth(float latitude_rad, float sd, float st,float altitude);
-float  solar_sunset(int month,int day);
+double solar_sunset(int month, int day);
 double sazi( double sd,  double st);	 /* solar azimuth from solar declination and solar time */
 double salt( double sd, double st);	   /* solar altitude from solar declination and solar time */
 double f_salt( double sd, double alt); /* solar altitude from solar declination and solar time */
